Fixing issue #25
As defined in issue #25, I was able to replicate an issue where git couldn't locate the path to the ssh client when cloning an org's private repos. In addition to this, I also had an issue with the ssh'ing without the known hosts. This fixes these problems I ran into.